Job Submitted Email

+++ aggregation-worker/src/main/java/au/org/emii/aggregationworker/AggregationRunner.java
@@ -1,45 +1,43 @@
 package au.org.emii.aggregationworker;
 
 
+import au.org.aodn.aws.exception.EmailException;
+import au.org.aodn.aws.util.EmailService;
+import au.org.aodn.aws.util.S3Utils;
 import au.org.aodn.aws.wps.status.EnumStatus;
 import au.org.aodn.aws.wps.status.ExecuteStatusBuilder;
 import au.org.aodn.aws.wps.status.S3StatusUpdater;
 import au.org.aodn.aws.wps.status.WpsConfig;
 import au.org.emii.aggregator.NetcdfAggregator;
-import au.org.emii.aggregator.overrides.AggregationOverrides;
+import au.org.emii.aggregator.catalogue.CatalogueReader;
 import au.org.emii.aggregator.converter.Converter;
-import au.org.emii.download.Download;
-import au.org.emii.download.DownloadConfig;
-import au.org.emii.download.DownloadRequest;
-import au.org.emii.download.Downloader;
-import au.org.emii.download.ParallelDownloadManager;
+import au.org.emii.aggregator.overrides.AggregationOverrides;
+import au.org.emii.download.*;
 import au.org.emii.geoserver.client.HttpIndexReader;
 import au.org.emii.geoserver.client.SubsetParameters;
-import au.org.emii.util.EmailService;
 import au.org.emii.util.IntegerHelper;
-import com.amazonaws.auth.DefaultAWSCredentialsProviderChain;
-import com.amazonaws.services.s3.transfer.TransferManager;
-import com.amazonaws.services.s3.transfer.Upload;
+import au.org.emii.util.ProvenanceWriter;
+import freemarker.template.Configuration;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
 import org.apache.commons.cli.DefaultParser;
 import org.apache.commons.cli.Options;
+import org.joda.time.DateTime;
+import org.joda.time.DateTimeZone;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
 import thredds.crawlabledataset.s3.S3URI;
-import ucar.nc2.time.CalendarDate;
 import ucar.nc2.time.CalendarDateRange;
-import ucar.unidata.geoloc.LatLonPoint;
-import ucar.unidata.geoloc.LatLonPointImmutable;
 import ucar.unidata.geoloc.LatLonRect;
+
 import java.io.UnsupportedEncodingException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.HashMap;
+import java.util.Map;
 import java.util.Set;
-
 
 import static au.org.aodn.aws.wps.status.WpsConfig.*;
 import static au.org.emii.aggregator.au.org.emii.aggregator.config.AggregationOverridesReader.getAggregationOverrides;
@@ -64,12 +62,11 @@ public class AggregationRunner implements CommandLineRunner {
     @Override
     public void run(String... args) {
 
+        DateTime startTime = new DateTime(DateTimeZone.UTC);
         S3StatusUpdater statusUpdater = null;
         String batchJobId = null, email = null;
         EmailService emailService = null;
-
-        String jobReportUrl = "jobReportUrl"; // Needed to be replaced
-        String expirationPeriod = "expirationPeriod"; // Needed to be replaced
+        ExecuteStatusBuilder statusBuilder = null;
 
         try {
             //  Capture the AWS job specifics - they are passed to the docker runtime as
@@ -84,35 +81,31 @@ public class AggregationRunner implements CommandLineRunner {
 
             String aggregatorConfigS3Bucket = WpsConfig.getConfig(AGGREGATOR_CONFIG_S3_BUCKET_CONFIG_KEY);
             String aggregatorTemplateFileS3Key = WpsConfig.getConfig(AGGREGATOR_TEMPLATE_FILE_S3_KEY_CONFIG_KEY);
+            String aggregatorProvenanceTemplateS3Key = WpsConfig.getConfig(PROVENANCE_TEMPLATE_S3_KEY_CONFIG_KEY);
 
             //  Parse connect timeout
             String downloadConnectTimeoutString = WpsConfig.getConfig(DOWNLOAD_CONNECT_TIMEOUT_CONFIG_KEY);
             int downloadConnectTimeout;
-            if(downloadConnectTimeoutString != null && IntegerHelper.isInteger(downloadConnectTimeoutString))
-            {
+            if (downloadConnectTimeoutString != null && IntegerHelper.isInteger(downloadConnectTimeoutString)) {
                 downloadConnectTimeout = Integer.parseInt(downloadConnectTimeoutString);
-            }
-            else
-            {
+            } else {
                 downloadConnectTimeout = DEFAULT_CONNECT_TIMEOUT_MS;
             }
 
             // Parse read timeout
             String downloadReadTimeoutString = WpsConfig.getConfig(DOWNLOAD_READ_TIMEOUT_CONFIG_KEY);
             int downloadReadTimeout;
-            if(downloadReadTimeoutString != null && IntegerHelper.isInteger(downloadReadTimeoutString))
-            {
+            if (downloadReadTimeoutString != null && IntegerHelper.isInteger(downloadReadTimeoutString)) {
                 downloadReadTimeout = Integer.parseInt(downloadConnectTimeoutString);
-            }
-            else
-            {
+            } else {
                 downloadReadTimeout = DEFAULT_READ_TIMEOUT_MS;
             }
 
 
             //  TODO:  null check and act on null configuration
             //  TODO : validate configuration
-            String statusDocument = ExecuteStatusBuilder.getStatusDocument(statusS3Bucket, statusFilename, batchJobId, EnumStatus.STARTED, null, null, null);
+            statusBuilder = new ExecuteStatusBuilder(batchJobId, statusS3Bucket, statusFilename);
+            String statusDocument = statusBuilder.createResponseDocument(EnumStatus.STARTED, null, null, null);
 
             //  Update status document to indicate job has started
             statusUpdater = new S3StatusUpdater(statusS3Bucket, statusFilename);
@@ -138,9 +131,9 @@ public class AggregationRunner implements CommandLineRunner {
             String resultMime = commandLine.getOptionValue('m');
             email = commandLine.getOptionValue('e');
 
-            SubsetParameters subsetParams = new SubsetParameters(subset);
+            SubsetParameters subsetParams = SubsetParameters.parse(subset);
 
-            if(email != null) {
+            if (email != null) {
                 email = email.substring(email.indexOf("=") + 1);
                 emailService = new EmailService();
             }
@@ -161,36 +154,16 @@ public class AggregationRunner implements CommandLineRunner {
             HttpIndexReader indexReader = new HttpIndexReader(WpsConfig.getConfig(WpsConfig.GEOSERVER_CATALOGUE_ENDPOINT_URL_CONFIG_KEY));
             Set<DownloadRequest> downloads = indexReader.getDownloadRequestList(layer, "time", "file_url", subsetParams);
 
-
-            LatLonRect bbox = null;
-            SubsetParameters.SubsetParameter latSubset = subsetParams.get("LATITUDE");
-            SubsetParameters.SubsetParameter lonSubset = subsetParams.get("LONGITUDE");
-
-            if (latSubset != null && lonSubset != null) {
-
-                double minLon = Double.parseDouble(lonSubset.start);
-                double minLat = Double.parseDouble(latSubset.start);
-                double maxLon = Double.parseDouble(lonSubset.end);
-                double maxLat = Double.parseDouble(latSubset.end);
-                LatLonPoint lowerLeft = new LatLonPointImmutable(minLat, minLon);
-                LatLonPoint upperRight = new LatLonPointImmutable(maxLat, maxLon);
-                bbox = new LatLonRect(lowerLeft, upperRight);
-
-                logger.info("Bounding box: LAT [" + minLat + ", " + maxLat + "], LON [" + minLon + ", " + maxLon + "]");
+            //  Apply subset parameters
+            LatLonRect bbox = subsetParams.getBbox();
+            if (bbox != null) {
+                logger.info("Bounding box: LAT [" + bbox.getLatMin() + ", " + bbox.getLatMax() + "], LON [" + bbox.getLonMin() + ", " + bbox.getLonMax() + "]");
             }
 
-
-            CalendarDateRange timeRange = null;
-
-            //  Apply time range (if provided)
-            SubsetParameters.SubsetParameter timeSubset = subsetParams.get("TIME");
-            if (timeSubset != null) {
-                CalendarDate startTime = CalendarDate.parseISOformat("Gregorian", timeSubset.start);
-                CalendarDate endTime = CalendarDate.parseISOformat("Gregorian", timeSubset.end);
-                timeRange = CalendarDateRange.of(startTime, endTime);
-                logger.info("Time range specified for aggregation: START [" + timeSubset.start + "], END [" + timeSubset.end + "]");
+            CalendarDateRange subsetTimeRange = subsetParams.getTimeRange();
+            if (subsetTimeRange != null) {
+                logger.info("Time range specified for aggregation: START [" + subsetTimeRange.getStart() + "], END [" + subsetTimeRange.getEnd() + "]");
             }
-
 
             //  Apply overrides (if provided)
             AggregationOverrides overrides = getAggregationOverrides(aggregatorConfigS3Bucket, aggregatorTemplateFileS3Key, null, layer);
@@ -208,7 +181,7 @@ public class AggregationRunner implements CommandLineRunner {
 
             try (
                     ParallelDownloadManager downloadManager = new ParallelDownloadManager(downloadConfig, downloader);
-                    NetcdfAggregator netcdfAggregator = new NetcdfAggregator(outputFile, overrides, chunkSize, bbox, null, timeRange)
+                    NetcdfAggregator netcdfAggregator = new NetcdfAggregator(outputFile, overrides, chunkSize, bbox, subsetParams.getVerticalRange(), subsetTimeRange)
             ) {
                 for (Download download : downloadManager.download(downloads)) {
                     netcdfAggregator.add(download.getPath());
@@ -224,40 +197,65 @@ public class AggregationRunner implements CommandLineRunner {
                 converter.convert(outputFile, convertedFile);
 
                 //  Form output file location
-                S3URI s3URI = new S3URI(outputBucketName, batchJobId + "/" + outputFilename + "." + converter.getExtension());
+                String jobPrefix = WpsConfig.getConfig(AWS_BATCH_JOB_S3_KEY);
+                S3URI resultS3URI = new S3URI(outputBucketName, jobPrefix + batchJobId + "/" + outputFilename + "." + converter.getExtension());
 
-                logger.info("Copying output file to : " + s3URI.toString());
+                logger.info("Copying output file to : " + resultS3URI.toString());
 
-                DefaultAWSCredentialsProviderChain credentialProviderChain = new DefaultAWSCredentialsProviderChain();
-                TransferManager tx = new TransferManager(credentialProviderChain.getCredentials());
-                Upload myUpload = tx.upload(s3URI.getBucket(), s3URI.getKey(), convertedFile.toFile());
-                myUpload.waitForCompletion();
-                tx.shutdownNow();
+                //  Upload to S3
+                S3Utils.uploadToS3(resultS3URI, convertedFile.toFile());
+
+                //  Lookup the metadata URL for the layer
+                String catalogueURL = WpsConfig.getConfig(GEONETWORK_CATALOGUE_URL_CONFIG_KEY);
+                String layerSearchField = WpsConfig.getConfig(GEONETWORK_CATALOGUE_LAYER_FIELD_CONFIG_KEY);
+                CatalogueReader catalogueReader = new CatalogueReader(catalogueURL, layerSearchField);
+
+                // Create provenance document
+                Configuration config = new Configuration();
+                config.setClassForTemplateLoading(ProvenanceWriter.class, "");
+
+                Map<String, Object> params = new HashMap<>();
+                params.put("jobId", batchJobId);
+                params.put("downloadUrl", WpsConfig.getS3ExternalURL(resultS3URI.getBucket(), resultS3URI.getKey()));
+                params.put("settingsPath", WpsConfig.getS3ExternalURL(aggregatorConfigS3Bucket, aggregatorTemplateFileS3Key));
+                params.put("startTime", startTime);
+                params.put("endTime", new DateTime(DateTimeZone.UTC));
+                params.put("layer", layer);
+                params.put("parameters", subsetParams);
+                params.put("sourceMetadataUrl", catalogueReader.getMetadataUrl(layer));
+                String provenanceDocument = ProvenanceWriter.write(aggregatorConfigS3Bucket, aggregatorProvenanceTemplateS3Key, params);
+
+                //  Upload provenance document to S3
+                //  TODO: configurable provenance filename?
+                S3URI provenanceS3URI = new S3URI(outputBucketName, jobPrefix + batchJobId + "/provenance.xml");
+                S3Utils.uploadToS3(provenanceS3URI, provenanceDocument);
 
                 HashMap<String, String> outputMap = new HashMap<>();
-                outputMap.put("result", WpsConfig.getS3ExternalURL(s3URI.getBucket(), s3URI.getKey()));
+                outputMap.put("result", WpsConfig.getS3ExternalURL(resultS3URI.getBucket(), resultS3URI.getKey()));
+                outputMap.put("provenance", WpsConfig.getS3ExternalURL(provenanceS3URI.getBucket(), provenanceS3URI.getKey()));
 
-                statusDocument = ExecuteStatusBuilder.getStatusDocument(statusS3Bucket, statusFilename, batchJobId, EnumStatus.SUCCEEDED, null, null, outputMap);
+                statusDocument = statusBuilder.createResponseDocument(EnumStatus.SUCCEEDED, null, null, outputMap);
                 statusUpdater.updateStatus(statusDocument, batchJobId);
             } finally {
-                Files.deleteIfExists(convertedFile);
-                Files.deleteIfExists(outputFile);
-            }
-
-            if(emailService != null) {
-                try {
-                    emailService.sendCompletedJobEmail(email, batchJobId, jobReportUrl, expirationPeriod);
-                } catch (Exception ex) {
-                    logger.error("Unable to send completed job email. Error Message:", ex);
+                if (convertedFile != null) {
+                    Files.deleteIfExists(convertedFile);
+                }
+                if (outputFile != null) {
+                    Files.deleteIfExists(outputFile);
                 }
             }
+
+            String jobPrefix = WpsConfig.getConfig(AWS_BATCH_JOB_S3_KEY);
+            String outputFileLocation = WpsConfig.getS3ExternalURL(outputBucketName, jobPrefix + batchJobId + "/" + outputFilename + "." + converter.getExtension());
+            emailService.sendCompletedJobEmail(email, batchJobId, outputFileLocation, WpsConfig.getJobExpiration());
+
         } catch (Throwable e) {
             e.printStackTrace();
             if (statusUpdater != null) {
                 if (batchJobId != null) {
                     String statusDocument = null;
                     try {
-                        statusDocument = ExecuteStatusBuilder.getStatusDocument(statusS3Bucket, statusFilename, batchJobId, EnumStatus.FAILED, "Exception occurred during aggregation :" + e.getMessage(), "AggregationError", null);
+                        statusDocument = statusBuilder.createResponseDocument(EnumStatus.FAILED, "Exception occurred during aggregation :" + e.getMessage(), "AggregationError", null);
                         statusUpdater.updateStatus(statusDocument, batchJobId);
                     } catch (UnsupportedEncodingException uex) {
                         logger.error("Unable to update status. Status: " + statusDocument);
@@ -266,11 +264,11 @@ public class AggregationRunner implements CommandLineRunner {
                 }
             }
 
-            if(emailService != null) {
+            if (emailService != null) {
                 try {
-                    emailService.sendFailedJobEmail(email, batchJobId, jobReportUrl);
-                } catch (Exception ex) {
-                    logger.error("Unable to send failed job email. Error Message:", ex);
+                    emailService.sendFailedJobEmail(email, batchJobId, "Job Report URL (Service not yet implemented)");
+                } catch (EmailException ex) {
+                    logger.error(ex.getMessage(), ex);
                 }
             }
             System.exit(1);
